+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 19.08
   * Adding feature to flush all requests in the queue
   * Adding CountlyConfig object which is used during init
+  * Added call to try to complete stored requests
+  * Merging Push module with SDK module
   * Fixing a potential 'null' bug for push
   * Fixing a consent bug which disabled features when consent was not enabled
   * Fixed a race condition bug when sending push token to server